+++ concordium-consensus/tests/globalstate/GlobalStateTests/Instances.hs
@@ -53,7 +53,7 @@ validContractArtifactsV1 :: [(Wasm.ModuleSource GSWasm.V1, GSWasm.ModuleInterfac
 validContractArtifactsV1 = mapMaybe packModule contractSourcesV1
     where packModule (_, sourceBytes) =
             let source = Wasm.ModuleSource sourceBytes
-            in (source,) <$> WasmV1.processModule (Wasm.WasmModuleV source)
+            in (source,) <$> WasmV1.processModule True (Wasm.WasmModuleV source)
 
 checkBinary :: Show a => (a -> a -> Bool) -> a -> a -> String -> String -> String -> Either String ()
 checkBinary bop x y sbop sx sy = unless (bop x y) $ Left $ "Not satisfied: " ++ sx ++ " (" ++ show x ++ ") " ++ sbop ++ " " ++ sy ++ " (" ++ show y ++ ")"
@@ -305,13 +305,13 @@ testUpdates n0 = if n0 <= 0 then return (property True) else tu n0 emptyInstance
                       InstanceDataV0 v a -> do
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                       InstanceDataV1 v a -> do
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                 updateExisting = do
@@ -323,7 +323,7 @@ testUpdates n0 = if n0 <= 0 then return (property True) else tu n0 emptyInstance
                         a <- arbitrary
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                       InstanceDataV1 _ _ -> do
@@ -331,7 +331,7 @@ testUpdates n0 = if n0 <= 0 then return (property True) else tu n0 emptyInstance
                         a <- arbitrary
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                 deleteExisting = do
@@ -349,13 +349,13 @@ testUpdates n0 = if n0 <= 0 then return (property True) else tu n0 emptyInstance
                       InstanceDataV0 v a -> do
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                       InstanceDataV1 v a -> do
                         let
                             ca = ContractAddress ci csi
-                            insts' = updateInstanceAt' ca a (Just v) insts
+                            insts' = updateInstanceAt' ca a (Just v) Nothing insts
                             model' = modelUpdateInstanceAt ca a v model
                         tu (n-1) insts' model'
                 deleteFree = do